longi ne rabotayut :(

+++ handlers.c
@@ -6,7 +6,7 @@
 /*   By: adavis <adavis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/03 14:35:39 by adavis            #+#    #+#             */
-/*   Updated: 2019/09/03 21:29:09 by adavis           ###   ########.fr       */
+/*   Updated: 2019/09/04 21:17:25 by adavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,9 +30,9 @@ int		handle_u(va_list ap, t_params *params)
 int		handle_d(va_list ap, t_params *params)
 {
 	if (params->l == 1)
-		return (d_render(va_arg(ap, long), params));
+		return (d_render((long)va_arg(ap, unsigned int), params));
 	else if (params->l == 2)
-		return (d_render(va_arg(ap, long long), params));
+		return (d_render((long)va_arg(ap, unsigned int), params));
 	else if (params->h == 1)
 		return (d_render((short)va_arg(ap, int), params));
 	else if (params->h == 2)
@@ -48,9 +48,11 @@ int		handle_x(va_list ap, t_params *params, t_bool upper)
 	else if (params->l == 2)
 		return (x_render(va_arg(ap, unsigned long long), params, upper));
 	else if (params->h == 1)
-		return (x_render((unsigned short int)va_arg(ap, unsigned int), params, upper));
+		return (x_render((unsigned short int)va_arg(ap, unsigned int), params,
+			upper));
 	else if (params->h == 2)
-		return (x_render((unsigned char)va_arg(ap, unsigned int), params, upper));
+		return (x_render((unsigned char)va_arg(ap, unsigned int), params,
+			upper));
 	else
 		return (x_render(va_arg(ap, unsigned int), params, upper));
 }
@@ -62,7 +64,7 @@ int		handle_o(va_list ap, t_params *params)
 	else if (params->l == 2)
 		return (o_render(va_arg(ap, unsigned long long), params));
 	else if (params->h == 1)
-		return (o_render((short int)va_arg(ap, int), params));
+		return (o_render((short)va_arg(ap, int), params));
 	else if (params->h == 2)
 		return (o_render((char)va_arg(ap, int), params));
 	else
@@ -72,7 +74,7 @@ int		handle_o(va_list ap, t_params *params)
 int		handle_f(va_list ap, t_params *params)
 {
 	if (params->precision == 0)
-		params->precision = 6; 
+		params->precision = 6;
 	if (params->l == 1)
 		return (f_render(va_arg(ap, double), params));
 	else if (params->lll)
